# api-bs-v1/src/main/resources/openapi/api-data-v1.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8082/api/data-v1
    description: Generated server url
paths:
  /persona:
    get:
      tags:
        - persona-controller
      operationId: getAll
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PersonaResponse'
    post:
      tags:
        - persona-controller
      operationId: save
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonaRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PersonaResponse'
  /persona/{id}:
    get:
      tags:
        - persona-controller
      operationId: findById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PersonaResponse'
    delete:
      tags:
        - persona-controller
      operationId: deleteById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
  /hello:
    get:
      tags:
        - hello-controller
      operationId: hello
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: string
components:
  schemas:
    PersonaRequest:
      type: object
      properties:
        id:
          type: integer
          format: int64
        documento:
          type: string
        nombre:
          type: string
        apellido:
          type: string
        edad:
          type: integer
          format: int32
        fechaNacimiento:
          type: string
    PersonaResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        documento:
          type: string
        nombre:
          type: string
        apellido:
          type: string
        edad:
          type: integer
          format: int32
        fechaNacimiento:
          type: string